+++ addons/main/controls/base.hpp
@@ -7,3 +7,17 @@ class Title: Default
         class Title;
     };
 };
+
+class Slider: Title
+{
+    class Controls;
+};
+class SliderMultiplier: Slider
+{
+    class Controls: Controls
+    {
+        class Title;
+        class Edit;
+        class Value;
+    };
+};